+++ lib/providers/badgeview_provider.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:badgemagic/badge_animation/ani_Animation.dart';
+import 'package:badgemagic/badge_animation/ani_animation.dart';
 import 'package:badgemagic/badge_animation/ani_down.dart';
 import 'package:badgemagic/badge_animation/ani_fixed.dart';
 import 'package:badgemagic/badge_animation/ani_laser.dart';
